backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "assistente-medico-api"
version = "0.1.0"
description = "API FastAPI do assistente médico (chat RAG PCDT)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.13",
    "aiosqlite>=0.20",
    "greenlet>=3.0",
    "sse-starlette>=2.0",
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-ollama>=0.2",
    "langchain-chroma>=0.1",
    "chromadb>=0.6",
    "grandalf>=0.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
